fix(resources): prevent circular import (resource -> detector -> resource -> ...)
