+++ packages/opentelemetry-resources/package.json
@@ -9,7 +9,11 @@
     "./src/platform/index.ts": "./src/platform/browser/index.ts",
     "./build/esm/platform/index.js": "./build/esm/platform/browser/index.js",
     "./build/esnext/platform/index.js": "./build/esnext/platform/browser/index.js",
-    "./build/src/platform/index.js": "./build/src/platform/browser/index.js"
+    "./build/src/platform/index.js": "./build/src/platform/browser/index.js",
+    "./src/detectors/platform/index.ts": "./src/detectors/platform/browser/index.ts",
+    "./build/esm/detectors/platform/index.js": "./build/esm/detectors/platform/browser/index.js",
+    "./build/esnext/detectors/platform/index.js": "./build/esnext/detectors/platform/browser/index.js",
+    "./build/src/detectors/platform/index.js": "./build/src/detectors/platform/browser/index.js"
   },
   "types": "build/src/index.d.ts",
   "repository": "open-telemetry/opentelemetry-js",
