+++ packages/opentelemetry-resources/src/detectors/platform/browser/OSDetector.ts
@@ -14,6 +14,6 @@
  * limitations under the License.
  */
 
-import { noopDetector } from '../../detectors/NoopDetector';
+import { noopDetector } from '../../NoopDetector';
 
 export const osDetector = noopDetector;
