+++ packages/opentelemetry-resources/src/detectors/platform/browser/ProcessDetector.ts
@@ -14,19 +14,6 @@
  * limitations under the License.
  */
 
-import { Detector } from '../../types';
-import { ResourceDetectionConfig } from '../../config';
-import { IResource } from '../../IResource';
-import { processDetectorSync } from './ProcessDetectorSync';
+import { noopDetector } from '../../NoopDetector';
 
-/**
- * ProcessDetector will be used to detect the resources related current process running
- * and being instrumented from the NodeJS Process module.
- */
-class ProcessDetector implements Detector {
-  detect(config?: ResourceDetectionConfig): Promise<IResource> {
-    return Promise.resolve(processDetectorSync.detect(config));
-  }
-}
-
-export const processDetector = new ProcessDetector();
+export const processDetector = noopDetector;
