+++ packages/opentelemetry-resources/src/detectors/platform/browser/ServiceInstanceIdDetectorSync.ts
@@ -14,6 +14,9 @@
  * limitations under the License.
  */
 
-import { noopDetector } from '../../detectors/NoopDetector';
+import { noopDetectorSync } from '../../NoopDetectorSync';
 
-export const processDetectorSync = noopDetector;
+/**
+ * @experimental
+ */
+export const serviceInstanceIdDetectorSync = noopDetectorSync;
