+++ packages/opentelemetry-resources/src/detectors/platform/browser/index.ts
@@ -14,6 +14,10 @@
  * limitations under the License.
  */
 
-import { noopDetector } from '../../detectors/NoopDetector';
-
-export const hostDetector = noopDetector;
+export { hostDetector } from './HostDetector';
+export { hostDetectorSync } from './HostDetectorSync';
+export { osDetector } from './OSDetector';
+export { osDetectorSync } from './OSDetectorSync';
+export { processDetector } from './ProcessDetector';
+export { processDetectorSync } from './ProcessDetectorSync';
+export { serviceInstanceIdDetectorSync } from './ServiceInstanceIdDetectorSync';
