+++ packages/opentelemetry-resources/src/detectors/platform/node/ProcessDetector.ts
@@ -14,6 +14,19 @@
  * limitations under the License.
  */
 
-import { noopDetectorSync } from '../../detectors/NoopDetectorSync';
+import { Detector } from '../../../types';
+import { ResourceDetectionConfig } from '../../../config';
+import { IResource } from '../../../IResource';
+import { processDetectorSync } from './ProcessDetectorSync';
 
-export const hostDetectorSync = noopDetectorSync;
+/**
+ * ProcessDetector will be used to detect the resources related current process running
+ * and being instrumented from the NodeJS Process module.
+ */
+class ProcessDetector implements Detector {
+  detect(config?: ResourceDetectionConfig): Promise<IResource> {
+    return Promise.resolve(processDetectorSync.detect(config));
+  }
+}
+
+export const processDetector = new ProcessDetector();
