+++ packages/opentelemetry-resources/src/detectors/platform/node/index.ts
@@ -14,22 +14,10 @@
  * limitations under the License.
  */
 
-export { Resource } from './Resource';
-export { IResource } from './IResource';
-export { defaultServiceName } from './platform';
-export { DetectorSync, ResourceAttributes, Detector } from './types';
-export { ResourceDetectionConfig } from './config';
-export {
-  browserDetector,
-  browserDetectorSync,
-  envDetector,
-  envDetectorSync,
-  hostDetector,
-  hostDetectorSync,
-  osDetector,
-  osDetectorSync,
-  processDetector,
-  processDetectorSync,
-  serviceInstanceIdDetectorSync,
-} from './detectors';
-export { detectResourcesSync, detectResources } from './detect-resources';
+export { hostDetector } from './HostDetector';
+export { hostDetectorSync } from './HostDetectorSync';
+export { osDetector } from './OSDetector';
+export { osDetectorSync } from './OSDetectorSync';
+export { processDetector } from './ProcessDetector';
+export { processDetectorSync } from './ProcessDetectorSync';
+export { serviceInstanceIdDetectorSync } from './ServiceInstanceIdDetectorSync';
