+++ packages/opentelemetry-resources/src/detectors/platform/node/machine-id/execAsync.ts
