+++ packages/opentelemetry-resources/src/detectors/platform/node/machine-id/getMachineId-darwin.ts
