+++ packages/opentelemetry-resources/src/detectors/platform/node/utils.ts
